# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LogB",
    "version": "0.1.0",
    "description": "The HTTP API behind LogB, served under `/api`.\n\nTwo credentials work everywhere except issuing and listing tokens: the `logb_session` cookie a browser gets from `POST /auth/login`, and an `Authorization: Bearer logb_pat_...` token for clients that are not browsers. A client obtains its first token by logging in interactively once and calling `POST /auth/tokens`; issuing and listing tokens deliberately refuse bearer credentials, so a leaked token cannot mint replacements. `DELETE /auth/tokens/{id}` is the one exception: a bearer token may revoke itself, so an app can sign out without ever holding a cookie -- but not any other token. `GET /health` lists `token-self-revoke` in its `features` array so a client can detect support before relying on it.\n\n**Writing offline.** `POST /objects/{id}/activities` and `POST /objects/{id}/attachments` accept a `client_op_id`: any string unique to the write, generated once by the client and kept across every retry. The server records it under a unique index, so replaying a write whose response was lost resolves to the row already created instead of duplicating it. Any client that queues writes while offline should use it.\n\nThis document is checked against the running router by `tests/openapi.rs`: a route added, removed or renamed without updating it fails the build."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "security": [
    {
      "sessionCookie": []
    },
    {
      "bearerToken": []
    }
  ],
  "paths": {
...
    "/stats": {
      "get": {
        "summary": "Spend across all of the caller's objects: over time, by object (children rolled up), by type and by category.",
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 9999
            },
            "description": "Restrict to one year; over_time then has twelve YYYY-MM buckets. Omitted: all years, one YYYY bucket each."
          },
          {
            "name": "purchases",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "Count objects' purchase prices, as by_category bucket purchase_price. Skipped for an object with a costed purchase entry."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Stats"
                }
              }
            }
          },
          "400": {
            "description": "year is not a number or outside 0..=9999"
          }
        }
      }
    },
    "/stats/energy": {
      "get": {
        "summary": "Household monthly electricity usage",
        "description": "Sums fuel quantities recorded in kWh across all owned, non-deleted objects. Months are oldest first and include the current partial month.",
        "parameters": [
          {
            "name": "months",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 36,
              "default": 12
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Household energy usage",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnergyUsage"
                }
              }
            }
          }
        }
      }
    },
    "/stats/fuel": {
      "get": {
        "summary": "Household monthly heating fuel usage",
        "description": "Sums fuel quantities in their recorded litres or gallons across all owned, non-deleted objects. No tank percentage is inferred because objects do not store tank capacity or level readings.",
        "parameters": [
          {
            "name": "months",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 36,
              "default": 12
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Household heating fuel usage",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FuelUsage"
                }
              }
            }
          }
        }
      }
    },
...
    "/stats/water": {
      "get": {
        "summary": "Household monthly water consumption",
        "description": "Combines billing-period usage with differences between cumulative meter readings. Volumes are normalized to thousandths of a litre.",
        "parameters": [
          {
            "name": "months",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 36,
              "default": 12
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Household water usage",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WaterUsage"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "logb_session",
        "description": "Set by `POST /auth/login`. HttpOnly, SameSite=Lax, 30 days."
      },
      "bearerToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "An API token from `POST /auth/tokens`. Does not expire; revoke it to end it."
      }
    },
    "schemas": {
...
      "StatsAmount": {
        "type": "object",
        "required": [
          "bucket",
          "cost_cents"
        ],
        "properties": {
          "bucket": {
            "type": "string"
          },
          "cost_cents": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "StatsObject": {
        "type": "object",
        "required": [
          "id",
          "name",
          "type",
          "archived",
          "cost_cents",
          "children"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "archived": {
            "type": "boolean"
          },
          "cost_cents": {
            "type": "integer",
            "format": "int64",
            "description": "Own spend plus every descendant's."
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StatsObject"
            }
          }
        }
      },
      "Stats": {
        "type": "object",
        "required": [
          "total_cents",
          "years",
          "over_time",
          "by_object",
          "by_type",
          "by_category"
        ],
        "properties": {
          "total_cents": {
            "type": "integer",
            "format": "int64"
          },
          "years": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Every year with spend, newest first; ignores the year filter."
          },
          "over_time": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StatsAmount"
            }
          },
          "by_object": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StatsObject"
            }
          },
          "by_type": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StatsAmount"
            }
          },
          "by_category": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StatsAmount"
            }
          }
        }
      },
      "EnergyUsage": {
        "type": "object",
        "required": [
          "months",
          "current_kwh_milli",
          "previous_kwh_milli",
          "target_kwh_milli"
        ],
        "properties": {
          "months": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EnergyUsageMonth"
            }
          },
          "current_kwh_milli": {
            "type": "integer",
            "format": "int64"
          },
          "previous_kwh_milli": {
            "type": "integer",
            "format": "int64"
          },
          "target_kwh_milli": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "EnergyUsageMonth": {
        "type": "object",
        "required": [
          "month",
          "kwh_milli",
          "charges",
          "objects"
        ],
        "properties": {
          "month": {
            "type": "string",
            "format": "month"
          },
          "kwh_milli": {
            "type": "integer",
            "format": "int64"
          },
          "charges": {
            "type": "integer",
            "format": "int64"
          },
          "objects": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "FuelUsage": {
        "type": "object",
        "required": [
          "months",
          "current_liters_milli",
          "previous_liters_milli",
          "current_gallons_milli",
          "previous_gallons_milli",
          "levels"
        ],
        "properties": {
          "months": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FuelUsageMonth"
            }
          },
          "current_liters_milli": {
            "type": "integer",
            "format": "int64"
          },
          "previous_liters_milli": {
            "type": "integer",
            "format": "int64"
          },
          "current_gallons_milli": {
            "type": "integer",
            "format": "int64"
          },
          "previous_gallons_milli": {
            "type": "integer",
            "format": "int64"
          },
          "levels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FuelTankLevel"
            }
          }
        }
      },
      "FuelTankLevel": {
        "type": "object",
        "required": [
          "object_id",
          "object_name",
          "unit",
          "date",
          "level_pct",
          "capacity_milli",
          "remaining_milli",
          "low",
          "estimated_days_remaining"
        ],
        "properties": {
          "object_id": {
            "type": "integer",
            "format": "int64"
          },
          "object_name": {
            "type": "string"
          },
          "unit": {
            "type": "string",
            "enum": [
              "l",
              "gal"
            ]
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "level_pct": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100
          },
          "capacity_milli": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "remaining_milli": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "low": {
            "type": "boolean"
          },
          "estimated_days_remaining": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        }
      },
      "FuelUsageMonth": {
        "type": "object",
        "required": [
          "month",
          "liters_milli",
          "gallons_milli",
          "charges",
          "objects"
        ],
        "properties": {
          "month": {
            "type": "string",
            "format": "month"
          },
          "liters_milli": {
            "type": "integer",
            "format": "int64"
          },
          "gallons_milli": {
            "type": "integer",
            "format": "int64"
          },
          "charges": {
            "type": "integer",
            "format": "int64"
          },
          "objects": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "WaterUsageMonth": {
        "type": "object",
        "required": [
          "month",
          "liters_milli",
          "cost_cents",
          "entries",
          "estimated"
        ],
        "properties": {
          "month": {
            "type": "string"
          },
          "liters_milli": {
            "type": "integer",
            "format": "int64"
          },
          "cost_cents": {
            "type": "integer",
            "format": "int64"
          },
          "entries": {
            "type": "integer",
            "format": "int64"
          },
          "estimated": {
            "type": "boolean"
          }
        }
      },
      "WaterObjectUsage": {
        "type": "object",
        "required": [
          "object_id",
          "object_name",
          "liters_milli",
          "target_liters_milli",
          "unit",
          "current_reading_milli"
        ],
        "properties": {
          "object_id": {
            "type": "integer",
            "format": "int64"
          },
          "object_name": {
            "type": "string"
          },
          "liters_milli": {
            "type": "integer",
            "format": "int64"
          },
          "target_liters_milli": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "unit": {
            "type": "string",
            "enum": [
              "l",
              "gal",
              "m3"
            ]
          },
          "current_reading_milli": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        }
      },
      "WaterUsage": {
        "type": "object",
        "required": [
          "months",
          "current_liters_milli",
          "previous_liters_milli",
          "current_cost_cents",
          "daily_average_liters_milli",
          "anomalies",
          "objects"
        ],
        "properties": {
          "months": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WaterUsageMonth"
            }
          },
          "current_liters_milli": {
            "type": "integer",
            "format": "int64"
          },
          "previous_liters_milli": {
            "type": "integer",
            "format": "int64"
          },
          "current_cost_cents": {
            "type": "integer",
            "format": "int64"
          },
          "daily_average_liters_milli": {
            "type": "integer",
            "format": "int64"
          },
          "anomalies": {
            "type": "integer",
            "format": "int64"
          },
          "objects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WaterObjectUsage"
            }
          }
        }
      }
    }
  }
}